rename dependencies package to platform
