+++ shared/src/androidMain/kotlin/com/maderskitech/kmpcodingexercisenetwork/platform/Platform.android.kt
@@ -1,4 +1,4 @@
-package com.maderskitech.kmpcodingexercisenetwork.dependencies
+package com.maderskitech.kmpcodingexercisenetwork.platform
 
 import android.os.Build
