+++ shared/src/commonMain/kotlin/com/maderskitech/kmpcodingexercisenetwork/data/remote/api/ItemsApi.kt
@@ -3,7 +3,7 @@ package com.maderskitech.kmpcodingexercisenetwork.data.remote.api
 import com.maderskitech.kmpcodingexercisenetwork.data.remote.api.model.ItemDto
 import com.maderskitech.kmpcodingexercisenetwork.data.remote.api.network.NetworkError
 import com.maderskitech.kmpcodingexercisenetwork.data.remote.api.network.Response
-import com.maderskitech.kmpcodingexercisenetwork.dependencies.NetworkClient
+import com.maderskitech.kmpcodingexercisenetwork.platform.NetworkClient
 import com.maderskitech.kmpcodingexercisenetwork.testingsupport.OpenForMokkery
 import io.ktor.client.call.body
 import io.ktor.client.request.get
