+++ shared/src/commonMain/kotlin/com/maderskitech/kmpcodingexercisenetwork/di/CommonModule.kt
@@ -4,6 +4,7 @@ import com.maderskitech.kmpcodingexercisenetwork.data.remote.api.ItemsApi
 import com.maderskitech.kmpcodingexercisenetwork.domain.mapper.ItemListMapper
 import com.maderskitech.kmpcodingexercisenetwork.domain.respository.DefaultItemRepository
 import com.maderskitech.kmpcodingexercisenetwork.domain.respository.ItemRepository
+import com.maderskitech.kmpcodingexercisenetwork.platform.di.platformModule
 import org.koin.core.module.dsl.factoryOf
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.bind
