+++ shared/src/commonMain/kotlin/com/maderskitech/kmpcodingexercisenetwork/platform/NetworkClient.kt
@@ -1,4 +1,4 @@
-package com.maderskitech.kmpcodingexercisenetwork.dependencies
+package com.maderskitech.kmpcodingexercisenetwork.platform
 
 import io.ktor.client.HttpClient
