+++ shared/src/iosMain/kotlin/com/maderskitech/kmpcodingexercisenetwork/platform/Platform.ios.kt
@@ -1,4 +1,4 @@
-package com.maderskitech.kmpcodingexercisenetwork.dependencies
+package com.maderskitech.kmpcodingexercisenetwork.platform
 
 import platform.UIKit.UIDevice
